feat(roll): roll Playwright to 1.13.0-next-1626254028000

+++ playwright/_impl/_api_structures.py
@@ -130,5 +130,5 @@ class SecurityDetails(TypedDict):
     issuer: Optional[str]
     protocol: Optional[str]
     subjectName: Optional[str]
-    validFrom: Optional[int]
-    validTo: Optional[int]
+    validFrom: Optional[float]
+    validTo: Optional[float]
